finish question 2
Changed the aproach order
